Remove ClassMetadataInfo
