+++ phpstan-baseline.neon
@@ -131,14 +131,14 @@ parameters:
 			path: lib/Doctrine/ORM/Mapping/ClassMetadataFactory.php
 
 		-
-			message: "#^Method Doctrine\\\\ORM\\\\Mapping\\\\ClassMetadataInfo\\:\\:fullyQualifiedClassName\\(\\) should return class\\-string\\|null but returns string\\|null\\.$#"
+			message: "#^Method Doctrine\\\\ORM\\\\Mapping\\\\ClassMetadata\\:\\:fullyQualifiedClassName\\(\\) should return class\\-string\\|null but returns string\\|null\\.$#"
 			count: 1
-			path: lib/Doctrine/ORM/Mapping/ClassMetadataInfo.php
+			path: lib/Doctrine/ORM/Mapping/ClassMetadata.php
 
 		-
 			message: "#^Negated boolean expression is always false\\.$#"
 			count: 1
-			path: lib/Doctrine/ORM/Mapping/ClassMetadataInfo.php
+			path: lib/Doctrine/ORM/Mapping/ClassMetadata.php
 
 		-
 			message: "#^Expression on left side of \\?\\? is not nullable\\.$#"
